+++ peer/bind.go
@@ -31,55 +31,59 @@ import (
 	"go.uber.org/multierr"
 )
 
-// Binder is a callback, provided to peer.Bind, that accepts a peer list and
-// binds it to a peer provider for the duration of the returned lifecycle.
-// The lifecycle that the binder returns should start and stop binding peers to
-// the list.
-// The binder must not call block on updating the list, because that will
-// typically block until the peer list has started.
-type Binder func(peer.List) transport.Lifecycle
-
-// Bind couples a peer list with a peer provider.
-// The peer provider must produce a binder that takes the peer list and returns
-// a lifecycle for the duration of the binding.
-func Bind(chooserList peer.ChooserList, bind Binder) *BoundChooser {
+// Bind couples a peer list with a peer list updater.
+// Bind accepts a peer list and passes that peer list to a binder.
+// The binder must return a peer list updater bound to the peer list.
+// The peer list updater must implement Lifecycle so the bound peer list
+// can start and stop updates.
+func Bind(chooserList peer.ChooserList, bind peer.Binder) *BoundChooser {
 	return &BoundChooser{
-		once:    intsync.Once(),
-		binding: bind(chooserList),
-		chooser: chooserList,
+		once:        intsync.Once(),
+		chooserList: chooserList,
+		updater:     bind(chooserList),
 	}
 }
 
-// BoundChooser is a peer chooser that couples a peer list and a peer provider
-// for the duration of its lifecycle.
+// BoundChooser is a peer chooser that couples a peer list and a peer list
+// updater for the duration of its lifecycle.
 type BoundChooser struct {
-	once    intsync.LifecycleOnce // the lifecycle of the bound peer chooser+provider
-	binding transport.Lifecycle   // the lifecycle of the peer provider
-	chooser peer.Chooser          // the peer chooser and the lifecycle of the peer chooser
+	once        intsync.LifecycleOnce
+	updater     transport.Lifecycle // the peer list updater, which to us is just a lifecycle
+	chooserList peer.ChooserList    // the peer list/chooser, also a lifecycle
+}
+
+// Updater returns the bound peer list updater.
+func (c *BoundChooser) Updater() transport.Lifecycle {
+	return c.updater
+}
+
+// ChooserList returns the bound peer list.
+func (c *BoundChooser) ChooserList() peer.ChooserList {
+	return c.chooserList
 }
 
 // Choose returns a peer from the bound peer list.
 func (c *BoundChooser) Choose(ctx context.Context, treq *transport.Request) (peer peer.Peer, onFinish func(error), err error) {
-	return c.chooser.Choose(ctx, treq)
+	return c.chooserList.Choose(ctx, treq)
 }
 
-// Start starts the peer list and the peer provider binding.
+// Start starts the peer list and the peer list updater.
 func (c *BoundChooser) Start() error {
 	return c.once.Start(c.start)
 }
 
 func (c *BoundChooser) start() error {
 
-	if err := c.chooser.Start(); err != nil {
+	if err := c.chooserList.Start(); err != nil {
 		return err
 	}
 
 	var errs error
-	if err := c.binding.Start(); err != nil {
+	if err := c.updater.Start(); err != nil {
 		errs = multierr.Append(errs, err)
 
-		// Abort the peer chooser if the binding failed to start.
-		if err := c.chooser.Stop(); err != nil {
+		// Abort the peer chooser if the updater failed to start.
+		if err := c.chooserList.Stop(); err != nil {
 			errs = multierr.Append(errs, err)
 		}
 	}
@@ -87,7 +91,7 @@ func (c *BoundChooser) start() error {
 	return errs
 }
 
-// Stop stops the peer list and the peer provider binding.
+// Stop stops the peer list and the peer list updater.
 func (c *BoundChooser) Stop() error {
 	return c.once.Stop(c.stop)
 }
@@ -95,27 +99,27 @@ func (c *BoundChooser) Stop() error {
 func (c *BoundChooser) stop() error {
 	var errs error
 
-	if err := c.binding.Stop(); err != nil {
+	if err := c.updater.Stop(); err != nil {
 		errs = multierr.Append(errs, err)
 	}
 
-	if err := c.chooser.Stop(); err != nil {
+	if err := c.chooserList.Stop(); err != nil {
 		errs = multierr.Append(errs, err)
 	}
 
 	return errs
 }
 
-// IsRunning returns whether the peer list and its peer provider binding are
-// both running, regardless of whether they should be running according to the
-// bound chooser's lifecycle.
+// IsRunning returns whether the peer list and its peer list updater are both
+// running, regardless of whether they should be running according to the bound
+// chooser's lifecycle.
 func (c *BoundChooser) IsRunning() bool {
-	return c.chooser.IsRunning() && c.binding.IsRunning()
+	return c.chooserList.IsRunning() && c.updater.IsRunning()
 }
 
 // Introspect introspects the bound chooser.
 func (c *BoundChooser) Introspect() introspection.ChooserStatus {
-	if ic, ok := c.chooser.(introspection.IntrospectableChooser); ok {
+	if ic, ok := c.chooserList.(introspection.IntrospectableChooser); ok {
 		return ic.Introspect()
 	}
 	return introspection.ChooserStatus{}
@@ -124,9 +128,9 @@ func (c *BoundChooser) Introspect() introspection.ChooserStatus {
 // BindPeers returns a binder (suitable as an argument to peer.Bind) that
 // binds a peer list to a static list of peers for the duration of its
 // lifecycle.
-func BindPeers(ids []peer.Identifier) Binder {
+func BindPeers(ids []peer.Identifier) peer.Binder {
 	return func(pl peer.List) transport.Lifecycle {
-		return &peersBinder{
+		return &PeersUpdater{
 			once: intsync.Once(),
 			pl:   pl,
 			ids:  ids,
@@ -134,32 +138,36 @@ func BindPeers(ids []peer.Identifier) Binder {
 	}
 }
 
-type peersBinder struct {
+// PeersUpdater binds a fixed list of peers to a peer list.
+type PeersUpdater struct {
 	once intsync.LifecycleOnce
 	pl   peer.List
 	ids  []peer.Identifier
 }
 
-func (s *peersBinder) Start() error {
+// Start adds a list of fixed peers to a peer list.
+func (s *PeersUpdater) Start() error {
 	return s.once.Start(s.start)
 }
 
-func (s *peersBinder) start() error {
+func (s *PeersUpdater) start() error {
 	return s.pl.Update(peer.ListUpdates{
 		Additions: s.ids,
 	})
 }
 
-func (s *peersBinder) Stop() error {
+// Stop removes a list of fixed peers from a peer list.
+func (s *PeersUpdater) Stop() error {
 	return s.once.Stop(s.stop)
 }
 
-func (s *peersBinder) stop() error {
+func (s *PeersUpdater) stop() error {
 	return s.pl.Update(peer.ListUpdates{
 		Removals: s.ids,
 	})
 }
 
-func (s *peersBinder) IsRunning() bool {
+// IsRunning returns whether the peers have been added and not removed.
+func (s *PeersUpdater) IsRunning() bool {
 	return s.once.IsRunning()
 }